take care of PR feedback.
