+++ lib/loader.js
@@ -4,10 +4,23 @@
 */
 var loaderUtils = require("loader-utils");
 var processCss = require("./processCss");
+var getExternals = require("./getExternals");
 var getImportPrefix = require("./getImportPrefix");
 var compileExports = require("./compile-exports");
 var createResolver = require("./createResolver");
 
+function findExternalFromImportUrl(importUrl, externals) {
+	var externalsKeys = Object.keys(externals);
+	var key = externalsKeys.find(function (externalsKey) {
+		return importUrl.endsWith(externalsKey);
+	});
+
+	if (key) {
+		return externals[key];
+	}
+
+	return false;
+}
 
 module.exports = function(content, map) {
 	if(this.cacheable) this.cacheable();
@@ -18,36 +31,7 @@ module.exports = function(content, map) {
 	var camelCaseKeys = query.camelCase || query.camelcase;
 	var sourceMap = query.sourceMap || false;
 	var resolve = createResolver(query.alias);
-
-	var givenExternals = query.externals || (this.externals || {});
-	var externals = {};
-	if (Array.isArray(givenExternals)) {
-		givenExternals.forEach(function (external) {
-			if (typeof external === 'object') {
-				externals = Object.assign({}, externals, external);
-			}
-		});
-	} else {
-		externals = givenExternals;
-	}
-
-	if (typeof externals === 'string') {
-		externals = JSON.parse(externals);
-	}
-
-	var externalsKeys = Object.keys(externals);
-
-	function findExternalFromImportUrl(importUrl) {
-		var key = externalsKeys.find(function (externalsKey) {
-			return importUrl.endsWith(externalsKey);
-		});
-
-		if (key) {
-			return externals[key];
-		}
-
-		return false;
-	}
+	var externals = getExternals(this, query);
 
 	if(sourceMap) {
 		if (map && typeof map !== "string") {
@@ -90,7 +74,7 @@ module.exports = function(content, map) {
 			} else {
 				var importUrl = importUrlPrefix + imp.url;
 
-				var external = findExternalFromImportUrl(importUrl);
+				var external = findExternalFromImportUrl(importUrl, externals);
 				if (external === false) {
 					return "exports.i(require(" + loaderUtils.stringifyRequest(this, importUrl) + "), " + JSON.stringify(imp.mediaQuery) + ");";
 				}
@@ -105,7 +89,7 @@ module.exports = function(content, map) {
 			var importItem = result.importItems[idx];
 			var importUrl = importUrlPrefix + importItem.url;
 
-			var external = findExternalFromImportUrl(importUrl);
+			var external = findExternalFromImportUrl(importUrl, externals);
 			if (external === false) {
 				return "\" + require(" + loaderUtils.stringifyRequest(this, importUrl) + ").locals" +
 				"[" + JSON.stringify(importItem.export) + "] + \"";
